# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-commerce API",
        "description": "API completa para o sistema de E-commerce Fullstack.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Servidor de Desenvolvimento"
        }
    ],
    "paths": {
        "/api/v1/addresses": {
            "get": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Ver endereço do usuário logado",
                "operationId": "39f13ca76f5e1930a55ae47a2bc11855",
                "responses": {
                    "200": {
                        "description": "Endereço retornado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "format": "uuid"
                                        },
                                        "street": {
                                            "type": "string"
                                        },
                                        "number": {
                                            "type": "integer"
                                        },
                                        "complement": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "neighborhood": {
                                            "type": "string"
                                        },
                                        "city": {
                                            "type": "string"
                                        },
                                        "state": {
                                            "type": "string"
                                        },
                                        "zip_code": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Endereço não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Atualizar endereço do usuário logado",
                "operationId": "8088de6c8442ec2684a0803ea535cdce",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "street": {
                                        "type": "string"
                                    },
                                    "number": {
                                        "type": "integer"
                                    },
                                    "complement": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "neighborhood": {
                                        "type": "string"
                                    },
                                    "city": {
                                        "type": "string"
                                    },
                                    "state": {
                                        "type": "string"
                                    },
                                    "zip_code": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Endereço atualizado com sucesso"
                    },
                    "404": {
                        "description": "Endereço não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Criar ou atualizar endereço do usuário logado",
                "operationId": "21aef37d3391058d7544becd47733c03",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "street",
                                    "number",
                                    "neighborhood",
                                    "city",
                                    "state",
                                    "zip_code"
                                ],
                                "properties": {
                                    "street": {
                                        "type": "string"
                                    },
                                    "number": {
                                        "type": "integer"
                                    },
                                    "complement": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "neighborhood": {
                                        "type": "string"
                                    },
                                    "city": {
                                        "type": "string"
                                    },
                                    "state": {
                                        "type": "string"
                                    },
                                    "zip_code": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Endereço salvo com sucesso"
                    },
                    "201": {
                        "description": "Endereço criado com sucesso"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Deletar endereço do usuário logado",
                "operationId": "befd723b176c3eb3d7d1a31e6d3627cf",
                "responses": {
                    "200": {
                        "description": "Endereço deletado com sucesso"
                    },
                    "404": {
                        "description": "Endereço não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}